+++ src/apps/content-editor/src/app/components/Editor/Editor.js
@@ -4,11 +4,6 @@ import cx from "classnames";
 import { AppLink } from "@zesty-io/core/AppLink";
 import { Breadcrumbs } from "shell/components/global-tabs/components/Breadcrumbs";
 import { Field } from "./Field";
-import { FieldV2 } from "./Field/FieldV2";
-import { FieldShell } from "./Field/FieldShell";
-import { withAI } from "../../../../../../shell/components/withAi";
-
-const AITextField = withAI(FieldShell);
 
 import styles from "./Editor.less";
 export default memo(function Editor({
